rename initMiscXXX in initSofaMiscXXX
