+++ modules/SofaMiscExtra/src/SofaMiscExtra/initSofaMiscExtra.cpp
@@ -19,21 +19,57 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_MISC_SOLVER_INIT_H
-#define SOFA_COMPONENT_MISC_SOLVER_INIT_H
-#include "config.h"
+#include <SofaMiscExtra/initSofaMiscExtra.h>
 
-namespace sofa
+#include <sofa/core/ObjectFactory.h>
+using sofa::core::ObjectFactory;
+
+namespace sofa::component
 {
 
-namespace component
+extern "C" {
+    SOFA_SOFAMISCEXTRA_API void initExternalModule();
+    SOFA_SOFAMISCEXTRA_API const char* getModuleName();
+    SOFA_SOFAMISCEXTRA_API const char* getModuleVersion();
+    SOFA_SOFAMISCEXTRA_API const char* getModuleLicense();
+    SOFA_SOFAMISCEXTRA_API const char* getModuleDescription();
+    SOFA_SOFAMISCEXTRA_API const char* getModuleComponentList();
+}
+
+void initExternalModule()
 {
+    static bool first = true;
+    if (first)
+    {
+        first = false;
+    }
+}
 
-void SOFA_MISC_SOLVER_API initMiscSolver();
+const char* getModuleName()
+{
+    return sofa_tostring(SOFA_TARGET);
+}
 
-} // namespace component
+const char* getModuleVersion()
+{
+    return sofa_tostring(SOFAMISCEXTRA_VERSION);
+}
 
-} // namespace sofa
+const char* getModuleLicense()
+{
+    return "LGPL";
+}
 
-#endif
+const char* getModuleDescription()
+{
+    return "This plugin contains contains features about Misc Extra.";
+}
 
+const char* getModuleComponentList()
+{
+    /// string containing the names of the classes provided by the plugin
+    static std::string classes = ObjectFactory::getInstance()->listClassesFromTarget(sofa_tostring(SOFA_TARGET));
+    return classes.c_str();
+}
+
+} // namespace sofa::component
