+++ modules/SofaMiscTopology/src/SofaMiscTopology/initSofaMiscTopology.cpp
@@ -19,22 +19,57 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_COMPONENT_MISC_FORCEFIELD_INIT_H
-#define SOFA_COMPONENT_MISC_FORCEFIELD_INIT_H
-#include "config.h"
+#include <SofaMiscTopology/initSofaMiscTopology.h>
 
-namespace sofa
+#include <sofa/core/ObjectFactory.h>
+using sofa::core::ObjectFactory;
+
+namespace sofa::component
 {
 
-namespace component
+extern "C" {
+    SOFA_SOFAMISCTOPOLOGY_API void initExternalModule();
+    SOFA_SOFAMISCTOPOLOGY_API const char* getModuleName();
+    SOFA_SOFAMISCTOPOLOGY_API const char* getModuleVersion();
+    SOFA_SOFAMISCTOPOLOGY_API const char* getModuleLicense();
+    SOFA_SOFAMISCTOPOLOGY_API const char* getModuleDescription();
+    SOFA_SOFAMISCTOPOLOGY_API const char* getModuleComponentList();
+}
+
+void initExternalModule()
 {
+    static bool first = true;
+    if (first)
+    {
+        first = false;
+    }
+}
 
+const char* getModuleName()
+{
+    return sofa_tostring(SOFA_TARGET);
+}
 
-void SOFA_MISC_FORCE_FIELD_API initMiscForcefield();
+const char* getModuleVersion()
+{
+    return sofa_tostring(SOFAMISCTOPOLOGY_VERSION);
+}
 
-} // namespace component
+const char* getModuleLicense()
+{
+    return "LGPL";
+}
 
-} // namespace sofa
+const char* getModuleDescription()
+{
+    return "This plugin contains contains features about Misc Topologies.";
+}
 
-#endif
+const char* getModuleComponentList()
+{
+    /// string containing the names of the classes provided by the plugin
+    static std::string classes = ObjectFactory::getInstance()->listClassesFromTarget(sofa_tostring(SOFA_TARGET));
+    return classes.c_str();
+}
 
+} // namespace sofa::component
